testutilsqemu: Always populate default ram id, default cpu, ACPI support for fake machines
Enable a few defaults that will decrease churn when converting tests to
real capabilities.

Since the fake machines are added only for x86_64 at this point we can
assume that ACPI is present via -machine.

In case of the default ram id assume the same. Additionally the logic
was broken as fake capabilities don't have a version so the default RAM
was never actually populated into fake caps tests.

For CPU we add 'qemu64' as that is the default picked by qemu.

We also assume that qemu requires an explicit backend for -numa, which
is the case for modern machines.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Reviewed-by: Ján Tomko <jtomko@redhat.com>